Move post methods from ApplicationHelper to PostHelper
